[Decomp][ObjectMgr] Cohesion split: extract locale/quest-relation/instance/spell/text loaders to ObjectMgr*.cpp
