+++ src/game/Object/ObjectMgr.cpp
@@ -316,251 +316,6 @@ ArenaTeam* ObjectMgr::GetArenaTeamByCaptain(ObjectGuid guid) const
     return NULL;
 }
 
-/**
- * @brief Loads localized creature names and subnames from the database.
- */
-void ObjectMgr::LoadCreatureLocales()
-{
-    mCreatureLocaleMap.clear();                             // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,"
-                                              "`name_loc1`,`subname_loc1`,`name_loc2`,`subname_loc2`,"
-                                              "`name_loc3`,`subname_loc3`,`name_loc4`,`subname_loc4`,"
-                                              "`name_loc5`,`subname_loc5`,`name_loc6`,`subname_loc6`,"
-                                              "`name_loc7`,`subname_loc7`,`name_loc8`,`subname_loc8`,"
-                                              "`name_loc9`,`subname_loc9`,`name_loc10`,`subname_loc10`,"
-                                              "`name_loc11`,`subname_loc11`"
-                                              " FROM `locales_creature`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 creature locale strings. DB table `locales_creature` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        if (!GetCreatureTemplate(entry))
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_creature` has data for not existed creature entry %u, skipped.", entry);
-            continue;
-        }
-
-        CreatureLocale& data = mCreatureLocaleMap[entry];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[1 + 2 * (i - 1)].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.Name.size() <= idx)
-                    {
-                        data.Name.resize(idx + 1);
-                    }
-
-                    data.Name[idx] = str;
-                }
-            }
-            str = fields[1 + 2 * (i - 1) + 1].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.SubName.size() <= idx)
-                    {
-                        data.SubName.resize(idx + 1);
-                    }
-
-                    data.SubName[idx] = str;
-                }
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu creature locale strings", mCreatureLocaleMap.size());
-    sLog.outString();
-}
-
-/**
- * @brief Loads localized gossip menu option and confirmation box text.
- */
-void ObjectMgr::LoadGossipMenuItemsLocales()
-{
-    mGossipMenuItemsLocaleMap.clear();                      // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `menu_id`,`id`,"
-                                              "`option_text_loc1`,`box_text_loc1`,`option_text_loc2`,`box_text_loc2`,"
-                                              "`option_text_loc3`,`box_text_loc3`,`option_text_loc4`,`box_text_loc4`,"
-                                              "`option_text_loc5`,`box_text_loc5`,`option_text_loc6`,`box_text_loc6`,"
-                                              "`option_text_loc7`,`box_text_loc7`,`option_text_loc8`,`box_text_loc8`,"
-                                              "`option_text_loc9`,`box_text_loc9`,`option_text_loc10`,`box_text_loc10`,"
-                                              "`option_text_loc11`,`box_text_loc11` "
-                                              "FROM `locales_gossip_menu_option`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString();
-        sLog.outString(">> Loaded 0 gossip_menu_option locale strings. DB table `locales_gossip_menu_option` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint16 menuId   = fields[0].GetUInt16();
-        uint16 id       = fields[1].GetUInt16();
-
-        GossipMenuItemsMapBounds bounds = GetGossipMenuItemsMapBounds(menuId);
-
-        bool found = false;
-        if (bounds.first != bounds.second)
-        {
-            for (GossipMenuItemsMap::const_iterator itr = bounds.first; itr != bounds.second; ++itr)
-            {
-                if (itr->second.id == id)
-                {
-                    found = true;
-                    break;
-                }
-            }
-        }
-
-        if (!found)
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_gossip_menu_option` has data for nonexistent gossip menu %u item %u, skipped.", menuId, id);
-            continue;
-        }
-
-        GossipMenuItemsLocale& data = mGossipMenuItemsLocaleMap[MAKE_PAIR32(menuId, id)];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[2 + 2 * (i - 1)].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.OptionText.size() <= idx)
-                    {
-                        data.OptionText.resize(idx + 1);
-                    }
-
-                    data.OptionText[idx] = str;
-                }
-            }
-            str = fields[2 + 2 * (i - 1) + 1].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.BoxText.size() <= idx)
-                    {
-                        data.BoxText.resize(idx + 1);
-                    }
-
-                    data.BoxText[idx] = str;
-                }
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu gossip_menu_option locale strings", mGossipMenuItemsLocaleMap.size());
-    sLog.outString();
-}
-
-/**
- * @brief Loads localized point-of-interest icon names.
- */
-void ObjectMgr::LoadPointOfInterestLocales()
-{
-    mPointOfInterestLocaleMap.clear();                      // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,"
-                                                "`icon_name_loc1`,`icon_name_loc2`,`icon_name_loc3`,`icon_name_loc4`,"
-                                                "`icon_name_loc5`,`icon_name_loc6`,`icon_name_loc7`,`icon_name_loc8`,"
-                                                "`icon_name_loc9`,`icon_name_loc10`,`icon_name_loc11`"
-                                                " FROM `locales_points_of_interest`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 points_of_interest locale strings. DB table `locales_points_of_interest` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        if (!GetPointOfInterest(entry))
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_points_of_interest` has data for nonexistent POI entry %u, skipped.", entry);
-            continue;
-        }
-
-        PointOfInterestLocale& data = mPointOfInterestLocaleMap[entry];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[i].GetCppString();
-            if (str.empty())
-            {
-                continue;
-            }
-
-            int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-            if (idx >= 0)
-            {
-                if ((int32)data.IconName.size() <= idx)
-                {
-                    data.IconName.resize(idx + 1);
-                }
-
-                data.IconName[idx] = str;
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu points_of_interest locale strings", mPointOfInterestLocaleMap.size());
-    sLog.outString();
-}
 
 // name must be checked to correctness (if received) before call this function
 ObjectGuid ObjectMgr::GetPlayerGuidByName(std::string name) const
@@ -767,436 +522,6 @@ CreatureDataAddon const* ObjectMgr::GetCreatureTemplateAddon(uint32 entry) { ret
  */
 ItemPrototype const* ObjectMgr::GetItemPrototype(uint32 id) { return sItemStorage.LookupEntry<ItemPrototype>(id); }
 
-/**
- * @brief Gets instance template data by map id.
- *
- * @param map The instance map id.
- * @return The instance template, or null if missing.
- */
-InstanceTemplate const* ObjectMgr::GetInstanceTemplate(uint32 map) { return sInstanceTemplate.LookupEntry<InstanceTemplate>(map); }
-WorldTemplate const* ObjectMgr::GetWorldTemplate(uint32 map) { return sWorldTemplate.LookupEntry<WorldTemplate>(map); }
-
-/* ********************************************************************************************* */
-/* *                                Loading Functions                                            */
-/* ********************************************************************************************* */
-void ObjectMgr::LoadArenaTeams()
-{
-    uint32 count = 0;
-
-    //                                                     0                          1      2             3      4                 5
-    QueryResult* result = CharacterDatabase.Query("SELECT `arena_team`.`arenateamid`,`name`,`captainguid`,`type`,`BackgroundColor`,`EmblemStyle`,"
-                          //   6          7             8              9        10           11          12             13            14
-                          "`EmblemColor`,`BorderStyle`,`BorderColor`, `rating`,`games_week`,`wins_week`,`games_season`,`wins_season`,`rank` "
-                          "FROM `arena_team` LEFT JOIN `arena_team_stats` ON `arena_team`.`arenateamid` = `arena_team_stats`.`arenateamid` ORDER BY `arena_team`.`arenateamid` ASC");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString();
-        sLog.outString(">> Loaded %u arenateam definitions", count);
-        return;
-    }
-
-    // load arena_team members
-    QueryResult* arenaTeamMembersResult = CharacterDatabase.Query(
-            //          0                   1      2             3           4               5             6                 7      8
-            "SELECT `arenateamid`,`member`.`guid`,`played_week`,`wons_week`,`played_season`,`wons_season`,`personal_rating`,`name`,`class` "
-            "FROM `arena_team_member` member LEFT JOIN `characters` chars on member.`guid` = chars.`guid` ORDER BY member.`arenateamid` ASC");
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-        ++count;
-
-        ArenaTeam* newArenaTeam = new ArenaTeam;
-        if (!newArenaTeam->LoadArenaTeamFromDB(result) ||
-                !newArenaTeam->LoadMembersFromDB(arenaTeamMembersResult))
-        {
-            newArenaTeam->Disband(NULL);
-            delete newArenaTeam;
-            continue;
-        }
-        AddArenaTeam(newArenaTeam);
-    }
-    while (result->NextRow());
-
-    delete result;
-    delete arenaTeamMembersResult;
-
-    sLog.outString();
-    sLog.outString(">> Loaded %u arenateam definitions", count);
-}
-
-/**
- * @brief Loads groups, group members, and group instance bindings from the database.
- */
-void ObjectMgr::LoadGroups()
-{
-    // -- loading groups --
-    uint32 count = 0;
-    //                                                     0           1                2             3             4                5        6        7        8        9        10       11       12       13           14            15                16            17
-    QueryResult* result = CharacterDatabase.Query("SELECT `mainTank`, `mainAssistant`, `lootMethod`, `looterGuid`, `lootThreshold`, `icon1`, `icon2`, `icon3`, `icon4`, `icon5`, `icon6`, `icon7`, `icon8`, `groupType`, `difficulty`, `raiddifficulty`, `leaderGuid`, `groupId` FROM `groups`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u group definitions", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-        ++count;
-        Group* group = new Group;
-        if (!group->LoadGroupFromDB(fields))
-        {
-            group->Disband();
-            delete group;
-            continue;
-        }
-        AddGroup(group);
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u group definitions", count);
-    sLog.outString();
-
-    // -- loading members --
-    count = 0;
-    //                                       0           1          2         3
-    result = CharacterDatabase.Query("SELECT `memberGuid`, `assistant`, `subgroup`, `groupId` FROM `group_member` ORDER BY `groupId`");
-    if (!result)
-    {
-        BarGoLink bar2(1);
-        bar2.step();
-    }
-    else
-    {
-        Group* group = NULL;                                // used as cached pointer for avoid relookup group for each member
-
-        BarGoLink bar2(result->GetRowCount());
-        do
-        {
-            bar2.step();
-            Field* fields = result->Fetch();
-            ++count;
-
-            uint32 memberGuidlow = fields[0].GetUInt32();
-            ObjectGuid memberGuid = ObjectGuid(HIGHGUID_PLAYER, memberGuidlow);
-            bool   assistent     = fields[1].GetBool();
-            uint8  subgroup      = fields[2].GetUInt8();
-            uint32 groupId       = fields[3].GetUInt32();
-            if (!group || group->GetId() != groupId)
-            {
-                group = GetGroupById(groupId);
-                if (!group)
-                {
-                    sLog.outErrorDb("Incorrect entry in group_member table : no group with Id %d for member %s!",
-                                    groupId, memberGuid.GetString().c_str());
-                    CharacterDatabase.PExecute("DELETE FROM `group_member` WHERE `memberGuid` = '%u'", memberGuidlow);
-                    continue;
-                }
-            }
-
-            if (!group->LoadMemberFromDB(memberGuidlow, subgroup, assistent))
-            {
-                sLog.outErrorDb("Incorrect entry in group_member table : member %s can not be added to group (Id: %u)!",
-                                memberGuid.GetString().c_str(), groupId);
-                CharacterDatabase.PExecute("DELETE FROM `group_member` WHERE `memberGuid` = '%u'", memberGuidlow);
-            }
-        }
-        while (result->NextRow());
-        delete result;
-    }
-
-    // clean groups
-    // TODO: maybe delete from the DB before loading in this case
-    for (GroupMap::iterator itr = mGroupMap.begin(); itr != mGroupMap.end();)
-    {
-        if (itr->second->GetMembersCount() < 2)
-        {
-            itr->second->Disband();
-            delete itr->second;
-            mGroupMap.erase(itr++);
-        }
-        else
-        {
-            ++itr;
-        }
-    }
-
-    // -- loading instances --
-    count = 0;
-    result = CharacterDatabase.Query(
-                 //                        0             1      2           3                       4             5
-                 "SELECT `group_instance`.`leaderGuid`, `map`, `instance`, `permanent`, `instance`.`difficulty`, `resettime`, "
-                 // 6
-                 "(SELECT COUNT(*) FROM `character_instance` WHERE `guid` = `group_instance`.`leaderGuid` AND `instance` = `group_instance`.`instance` AND `permanent` = 1 LIMIT 1), "
-                 // 7                              8
-                 " `groups`.`groupId`, `instance`.`encountersMask` "
-                 "FROM `group_instance` LEFT JOIN `instance` ON `instance` = `id` LEFT JOIN `groups` ON `groups`.`leaderGUID` = `group_instance`.`leaderGUID` ORDER BY `leaderGuid`"
-             );
-
-    if (!result)
-    {
-        BarGoLink bar2(1);
-        bar2.step();
-    }
-    else
-    {
-        Group* group = NULL;                                // used as cached pointer for avoid relookup group for each member
-
-        BarGoLink bar2(result->GetRowCount());
-        do
-        {
-            bar2.step();
-            Field* fields = result->Fetch();
-            ++count;
-
-            uint32 leaderGuidLow = fields[0].GetUInt32();
-            uint32 mapId = fields[1].GetUInt32();
-            Difficulty diff = (Difficulty)fields[4].GetUInt8();
-            uint32 groupId = fields[7].GetUInt32();
-
-            if (!group || group->GetId() != groupId)
-            {
-                // find group id in map by leader low guid
-                group = GetGroupById(groupId);
-                if (!group)
-                {
-                    sLog.outErrorDb("Incorrect entry in group_instance table : no group with leader %d", leaderGuidLow);
-                    continue;
-                }
-            }
-
-            MapEntry const* mapEntry = sMapStore.LookupEntry(mapId);
-            if (!mapEntry || !mapEntry->IsDungeon())
-            {
-                sLog.outErrorDb("Incorrect entry in group_instance table : no dungeon map %d", mapId);
-                continue;
-            }
-
-            if (diff >= (mapEntry->IsRaid() ? MAX_RAID_DIFFICULTY : MAX_DUNGEON_DIFFICULTY))
-            {
-                sLog.outErrorDb("Wrong dungeon difficulty use in group_instance table: %d", diff + 1);
-                diff = REGULAR_DIFFICULTY;                  // default for both difficaly types
-            }
-
-            DungeonPersistentState* state = (DungeonPersistentState*)sMapPersistentStateMgr.AddPersistentState(mapEntry, fields[2].GetUInt32(), Difficulty(diff), (time_t)fields[5].GetUInt64(), (fields[6].GetUInt32() == 0), true, true, fields[8].GetUInt32());
-            group->BindToInstance(state, fields[3].GetBool(), true);
-        }
-        while (result->NextRow());
-        delete result;
-    }
-
-    sLog.outString(">> Loaded %u group-instance binds total", count);
-    sLog.outString();
-
-    sLog.outString(">> Loaded %u group members total", count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads page text records and validates page chains.
- */
-void ObjectMgr::LoadPageTexts()
-{
-    sPageTextStore.Load();
-    sLog.outString(">> Loaded %u page texts", sPageTextStore.GetRecordCount());
-    sLog.outString();
-
-    for (uint32 i = 1; i < sPageTextStore.GetMaxEntry(); ++i)
-    {
-        // check data correctness
-        PageText const* page = sPageTextStore.LookupEntry<PageText>(i);
-        if (!page)
-        {
-            continue;
-        }
-
-        if (page->Next_Page && !sPageTextStore.LookupEntry<PageText>(page->Next_Page))
-        {
-            sLog.outErrorDb("Page text (Id: %u) has not existing next page (Id:%u)", i, page->Next_Page);
-            continue;
-        }
-
-        // detect circular reference
-        std::set<uint32> checkedPages;
-        for (PageText const* pageItr = page; pageItr; pageItr = sPageTextStore.LookupEntry<PageText>(pageItr->Next_Page))
-        {
-            if (!pageItr->Next_Page)
-            {
-                break;
-            }
-            checkedPages.insert(pageItr->Page_ID);
-            if (checkedPages.find(pageItr->Next_Page) != checkedPages.end())
-            {
-                std::ostringstream ss;
-                ss << "The text page(s) ";
-                for (std::set<uint32>::iterator itr = checkedPages.begin(); itr != checkedPages.end(); ++itr)
-                {
-                    ss << *itr << " ";
-                }
-                ss << "create(s) a circular reference, which can cause the server to freeze. Changing Next_Page of page "
-                   << pageItr->Page_ID << " to 0";
-                sLog.outErrorDb("%s", ss.str().c_str());
-                const_cast<PageText*>(pageItr)->Next_Page = 0;
-                break;
-            }
-        }
-    }
-}
-
-/**
- * @brief Loads localized page text content.
- */
-void ObjectMgr::LoadPageTextLocales()
-{
-    mPageTextLocaleMap.clear();                             // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,"
-                                              "`text_loc1`,`text_loc2`,`text_loc3`,`text_loc4`,"
-                                              "`text_loc5`,`text_loc6`,`text_loc7`,`text_loc8`,"
-                                              "`text_loc9`,`text_loc10`,`text_loc11`"
-                                              " FROM `locales_page_text`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 PageText locale strings. DB table `locales_page_text` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        if (!sPageTextStore.LookupEntry<PageText>(entry))
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_page_text` has data for nonexistent page text entry %u, skipped.", entry);
-            continue;
-        }
-
-        PageTextLocale& data = mPageTextLocaleMap[entry];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[i].GetCppString();
-            if (str.empty())
-            {
-                continue;
-            }
-
-            int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-            if (idx >= 0)
-            {
-                if ((int32)data.Text.size() <= idx)
-                {
-                    data.Text.resize(idx + 1);
-                }
-
-                data.Text[idx] = str;
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu PageText locale strings", mPageTextLocaleMap.size());
-    sLog.outString();
-}
-
-void ObjectMgr::LoadInstanceEncounters()
-{
-    m_DungeonEncounters.clear();         // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `creditType`, `creditEntry`, `lastEncounterDungeon` FROM `instance_encounters`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString();
-        sLog.outString(">> Loaded 0 Instance Encounters. DB table `instance_encounters` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-        DungeonEncounterEntry const* dungeonEncounter = sDungeonEncounterStore.LookupEntry(entry);
-
-        if (!dungeonEncounter)
-        {
-            sLog.outErrorDb("Table `instance_encounters` has an invalid encounter id %u, skipped!", entry);
-            continue;
-        }
-
-        uint8 creditType = fields[1].GetUInt8();
-        uint32 creditEntry = fields[2].GetUInt32();
-        switch (creditType)
-        {
-            case ENCOUNTER_CREDIT_KILL_CREATURE:
-            {
-                CreatureInfo const* cInfo = sCreatureStorage.LookupEntry<CreatureInfo>(creditEntry);
-                if (!cInfo)
-                {
-                    sLog.outErrorDb("Table `instance_encounters` has an invalid creature (entry %u) linked to the encounter %u (%s), skipped!", creditEntry, entry, dungeonEncounter->encounterName[0]);
-                    continue;
-                }
-                break;
-            }
-            case ENCOUNTER_CREDIT_CAST_SPELL:
-            {
-                if (!sSpellStore.LookupEntry(creditEntry))
-                {
-                    // skip spells that aren't in dbc for now
-                    // sLog.outErrorDb("Table `instance_encounters` has an invalid spell (entry %u) linked to the encounter %u (%s), skipped!", creditEntry, entry, dungeonEncounter->encounterName[0]);
-                    continue;
-                }
-                break;
-            }
-            default:
-                sLog.outErrorDb("Table `instance_encounters` has an invalid credit type (%u) for encounter %u (%s), skipped!", creditType, entry, dungeonEncounter->encounterName[0]);
-                continue;
-        }
-        uint32 lastEncounterDungeon = fields[3].GetUInt32();
-
-        m_DungeonEncounters.insert(DungeonEncounterMap::value_type(creditEntry, new DungeonEncounter(dungeonEncounter, EncounterCreditType(creditType), creditEntry, lastEncounterDungeon)));
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString();
-    sLog.outString(">> Loaded %zu Instance Encounters", m_DungeonEncounters.size());
-}
 
 struct SQLInstanceLoader : public SQLStorageLoaderBase<SQLInstanceLoader, SQLStorage>
 {
@@ -1350,153 +675,6 @@ GossipText const* ObjectMgr::GetGossipText(uint32 Text_ID) const
     return NULL;
 }
 
-/**
- * @brief Loads npc gossip text records from the database.
- */
-void ObjectMgr::LoadGossipText()
-{
-    QueryResult* result = WorldDatabase.Query("SELECT * FROM `npc_text`");
-
-    int count = 0;
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString(">> Loaded %u npc texts", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        ++count;
-        int cic = 0;
-
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 Text_ID    = fields[cic++].GetUInt32();
-        if (!Text_ID)
-        {
-            sLog.outErrorDb("Table `npc_text` has record wit reserved id 0, ignore.");
-            continue;
-        }
-
-        GossipText& gText = mGossipText[Text_ID];
-
-        for (int i = 0; i < MAX_GOSSIP_TEXT_OPTIONS; ++i)
-        {
-            gText.Options[i].Text_0           = fields[cic++].GetCppString();
-            gText.Options[i].Text_1           = fields[cic++].GetCppString();
-
-            gText.Options[i].Language         = fields[cic++].GetUInt32();
-            gText.Options[i].Probability      = fields[cic++].GetFloat();
-
-            for (int j = 0; j < 3; ++j)
-            {
-                gText.Options[i].Emotes[j]._Delay  = fields[cic++].GetUInt32();
-                gText.Options[i].Emotes[j]._Emote  = fields[cic++].GetUInt32();
-            }
-        }
-    }
-    while (result->NextRow());
-
-    sLog.outString(">> Loaded %u npc texts", count);
-    sLog.outString();
-    delete result;
-}
-
-/**
- * @brief Loads localized npc gossip text variants.
- */
-void ObjectMgr::LoadGossipTextLocales()
-{
-    mNpcTextLocaleMap.clear();                              // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,"
-                                              "`Text0_0_loc1`,`Text0_1_loc1`,`Text1_0_loc1`,`Text1_1_loc1`,`Text2_0_loc1`,`Text2_1_loc1`,`Text3_0_loc1`,`Text3_1_loc1`,`Text4_0_loc1`,`Text4_1_loc1`,`Text5_0_loc1`,`Text5_1_loc1`,`Text6_0_loc1`,`Text6_1_loc1`,`Text7_0_loc1`,`Text7_1_loc1`,"
-                                              "`Text0_0_loc2`,`Text0_1_loc2`,`Text1_0_loc2`,`Text1_1_loc2`,`Text2_0_loc2`,`Text2_1_loc2`,`Text3_0_loc2`,`Text3_1_loc2`,`Text4_0_loc2`,`Text4_1_loc2`,`Text5_0_loc2`,`Text5_1_loc2`,`Text6_0_loc2`,`Text6_1_loc2`,`Text7_0_loc2`,`Text7_1_loc2`,"
-                                              "`Text0_0_loc3`,`Text0_1_loc3`,`Text1_0_loc3`,`Text1_1_loc3`,`Text2_0_loc3`,`Text2_1_loc3`,`Text3_0_loc3`,`Text3_1_loc3`,`Text4_0_loc3`,`Text4_1_loc3`,`Text5_0_loc3`,`Text5_1_loc3`,`Text6_0_loc3`,`Text6_1_loc3`,`Text7_0_loc3`,`Text7_1_loc3`,"
-                                              "`Text0_0_loc4`,`Text0_1_loc4`,`Text1_0_loc4`,`Text1_1_loc4`,`Text2_0_loc4`,`Text2_1_loc4`,`Text3_0_loc4`,`Text3_1_loc4`,`Text4_0_loc4`,`Text4_1_loc4`,`Text5_0_loc4`,`Text5_1_loc4`,`Text6_0_loc4`,`Text6_1_loc4`,`Text7_0_loc4`,`Text7_1_loc4`,"
-                                              "`Text0_0_loc5`,`Text0_1_loc5`,`Text1_0_loc5`,`Text1_1_loc5`,`Text2_0_loc5`,`Text2_1_loc5`,`Text3_0_loc5`,`Text3_1_loc5`,`Text4_0_loc5`,`Text4_1_loc5`,`Text5_0_loc5`,`Text5_1_loc5`,`Text6_0_loc5`,`Text6_1_loc5`,`Text7_0_loc5`,`Text7_1_loc5`,"
-                                              "`Text0_0_loc6`,`Text0_1_loc6`,`Text1_0_loc6`,`Text1_1_loc6`,`Text2_0_loc6`,`Text2_1_loc6`,`Text3_0_loc6`,`Text3_1_loc6`,`Text4_0_loc6`,`Text4_1_loc6`,`Text5_0_loc6`,`Text5_1_loc6`,`Text6_0_loc6`,`Text6_1_loc6`,`Text7_0_loc6`,`Text7_1_loc6`,"
-                                              "`Text0_0_loc7`,`Text0_1_loc7`,`Text1_0_loc7`,`Text1_1_loc7`,`Text2_0_loc7`,`Text2_1_loc7`,`Text3_0_loc7`,`Text3_1_loc7`,`Text4_0_loc7`,`Text4_1_loc7`,`Text5_0_loc7`,`Text5_1_loc7`,`Text6_0_loc7`,`Text6_1_loc7`,`Text7_0_loc7`,`Text7_1_loc7`, "
-                                              "`Text0_0_loc8`,`Text0_1_loc8`,`Text1_0_loc8`,`Text1_1_loc8`,`Text2_0_loc8`,`Text2_1_loc8`,`Text3_0_loc8`,`Text3_1_loc8`,`Text4_0_loc8`,`Text4_1_loc8`,`Text5_0_loc8`,`Text5_1_loc8`,`Text6_0_loc8`,`Text6_1_loc8`,`Text7_0_loc8`,`Text7_1_loc8`,"
-                                              "`Text0_0_loc9`,`Text0_1_loc9`,`Text1_0_loc9`,`Text1_1_loc9`,`Text2_0_loc9`,`Text2_1_loc9`,`Text3_0_loc9`,`Text3_1_loc9`,`Text4_0_loc9`,`Text4_1_loc9`,`Text5_0_loc9`,`Text5_1_loc9`,`Text6_0_loc9`,`Text6_1_loc9`,`Text7_0_loc9`,`Text7_1_loc9`,"
-                                              "`Text0_0_loc10`,`Text0_1_loc10`,`Text1_0_loc10`,`Text1_1_loc10`,`Text2_0_loc10`,`Text2_1_loc10`,`Text3_0_loc10`,`Text3_1_loc10`,`Text4_0_loc10`,`Text4_1_loc10`,`Text5_0_loc10`,`Text5_1_loc10`,`Text6_0_loc10`,`Text6_1_loc10`,`Text7_0_loc10`,`Text7_1_loc10`,"
-                                              "`Text0_0_loc11`,`Text0_1_loc11`,`Text1_0_loc11`,`Text1_1_loc11`,`Text2_0_loc11`,`Text2_1_loc11`,`Text3_0_loc11`,`Text3_1_loc11`,`Text4_0_loc11`,`Text4_1_loc11`,`Text5_0_loc11`,`Text5_1_loc11`,`Text6_0_loc11`,`Text6_1_loc11`,`Text7_0_loc11`,`Text7_1_loc11` "
-                                              " FROM `locales_npc_text`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 Quest locale strings. DB table `locales_npc_text` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        if (!GetGossipText(entry))
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_npc_text` has data for nonexistent gossip text entry %u, skipped.", entry);
-            continue;
-        }
-
-        NpcTextLocale& data = mNpcTextLocaleMap[entry];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            for (int j = 0; j < 8; ++j)
-            {
-                std::string str0 = fields[1 + 8 * 2 * (i - 1) + 2 * j].GetCppString();
-                if (!str0.empty())
-                {
-                    int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                    if (idx >= 0)
-                    {
-                        if ((int32)data.Text_0[j].size() <= idx)
-                        {
-                            data.Text_0[j].resize(idx + 1);
-                        }
-
-                        data.Text_0[j][idx] = str0;
-                    }
-                }
-                std::string str1 = fields[1 + 8 * 2 * (i - 1) + 2 * j + 1].GetCppString();
-                if (!str1.empty())
-                {
-                    int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                    if (idx >= 0)
-                    {
-                        if ((int32)data.Text_1[j].size() <= idx)
-                        {
-                            data.Text_1[j].resize(idx + 1);
-                        }
-
-                        data.Text_1[j][idx] = str1;
-                    }
-                }
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu NpcText locale strings", mNpcTextLocaleMap.size());
-    sLog.outString();
-}
 
 // not very fast function but it is called only once a day, or on starting-up
 /// @param serverUp true if the server is already running, false when the server is started
@@ -1899,92 +1077,6 @@ void ObjectMgr::SetHighestGuids()
     m_FirstTemporaryGameObjectGuid += sWorld.getConfig(CONFIG_UINT32_GUID_RESERVE_SIZE_GAMEOBJECT);
 }
 
-/**
- * @brief Loads localized gameobject names.
- */
-void ObjectMgr::LoadGameObjectLocales()
-{
-    mGameObjectLocaleMap.clear();                           // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,"
-                                              "`name_loc1`,`name_loc2`,`name_loc3`,`name_loc4`,"
-                                              "`name_loc5`,`name_loc6`,`name_loc7`,`name_loc8`,"
-                                              "`name_loc9`,`name_loc10`,`name_loc11`,"
-                                              "`castbarcaption_loc1`,`castbarcaption_loc2`,"
-                                              "`castbarcaption_loc3`,`castbarcaption_loc4`,"
-                                              "`castbarcaption_loc5`,`castbarcaption_loc6`,"
-                                              "`castbarcaption_loc7`,`castbarcaption_loc8`,"
-                                              "`castbarcaption_loc9`,`castbarcaption_loc10`,"
-                                              "`castbarcaption_loc11`"
-                                              " FROM `locales_gameobject`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 gameobject locale strings. DB table `locales_gameobject` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        if (!GetGameObjectInfo(entry))
-        {
-            ERROR_DB_STRICT_LOG("Table `locales_gameobject` has data for nonexistent gameobject entry %u, skipped.", entry);
-            continue;
-        }
-
-        GameObjectLocale& data = mGameObjectLocaleMap[entry];
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[i].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.Name.size() <= idx)
-                    {
-                        data.Name.resize(idx + 1);
-                    }
-
-                    data.Name[idx] = str;
-                }
-            }
-        }
-
-        for (int i = 1; i < MAX_LOCALE; ++i)
-        {
-            std::string str = fields[i + (MAX_LOCALE - 1)].GetCppString();
-            if (!str.empty())
-            {
-                int idx = GetOrNewIndexForLocale(LocaleConstant(i));
-                if (idx >= 0)
-                {
-                    if ((int32)data.CastBarCaption.size() <= idx)
-                    {
-                        data.CastBarCaption.resize(idx + 1);
-                    }
-
-                    data.CastBarCaption[idx] = str;
-                }
-            }
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %zu gameobject locale strings", mGameObjectLocaleMap.size());
-    sLog.outString();
-}
 
 /**
  * @brief Loads exploration base experience values by level.
@@ -2326,95 +1418,6 @@ void ObjectMgr::LoadQuestPOI()
     sLog.outString(">> Loaded %u quest POI definitions", count);
 }
 
-void ObjectMgr::LoadNPCSpellClickSpells()
-{
-    uint32 count = 0;
-
-    mSpellClickInfoMap.clear();
-    //                                                 0            1           2              3                     4            5             6
-    QueryResult* result = WorldDatabase.Query("SELECT `npc_entry`, `spell_id`, `quest_start`, `quest_start_active`, `quest_end`, `cast_flags`, `condition_id` FROM `npc_spellclick_spells`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString();
-        sLog.outErrorDb(">> Loaded 0 spellclick spells. DB table `npc_spellclick_spells` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        SpellClickInfo info;
-        uint32 npc_entry         = fields[0].GetUInt32();
-        info.spellId             = fields[1].GetUInt32();
-        info.questStart          = fields[2].GetUInt32();
-        info.questStartCanActive = fields[3].GetBool();
-        info.questEnd            = fields[4].GetUInt32();
-        info.castFlags           = fields[5].GetUInt8();
-        info.conditionId         = fields[6].GetUInt16();
-
-        CreatureInfo const* cInfo = GetCreatureTemplate(npc_entry);
-        if (!cInfo)
-        {
-            sLog.outErrorDb("Table npc_spellclick_spells references unknown creature_template %u. Skipping entry.", npc_entry);
-            continue;
-        }
-
-        // spell can be 0 for special or custom cases
-        if (info.spellId)
-        {
-            SpellEntry const* spellinfo = sSpellStore.LookupEntry(info.spellId);
-            if (!spellinfo)
-            {
-                sLog.outErrorDb("Table npc_spellclick_spells references unknown spellid %u. Skipping entry.", info.spellId);
-                continue;
-            }
-        }
-
-        if (info.conditionId && !sConditionStorage.LookupEntry<PlayerCondition const*>(info.conditionId))
-        {
-            sLog.outErrorDb("Table npc_spellclick_spells references unknown condition %u. Skipping entry.", info.conditionId);
-            continue;
-        }
-        else if (!info.conditionId)                         // TODO Drop block after finished converting
-        {
-            // quest might be 0 to enable spellclick independent of any quest
-            if (info.questStart && mQuestTemplates.find(info.questStart) == mQuestTemplates.end())
-            {
-                sLog.outErrorDb("Table npc_spellclick_spells references unknown start quest %u. Skipping entry.", info.questStart);
-                continue;
-            }
-
-            // quest might be 0 to enable spellclick active infinity after start quest
-            if (info.questEnd && mQuestTemplates.find(info.questEnd) == mQuestTemplates.end())
-            {
-                sLog.outErrorDb("Table npc_spellclick_spells references unknown end quest %u. Skipping entry.", info.questEnd);
-                continue;
-            }
-        }
-
-        mSpellClickInfoMap.insert(SpellClickInfoMap::value_type(npc_entry, info));
-
-        // mark creature template as spell clickable
-        const_cast<CreatureInfo*>(cInfo)->NpcFlags |= UNIT_NPC_FLAG_SPELLCLICK;
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString();
-    sLog.outString(">> Loaded %u spellclick definitions", count);
-}
 
 void ObjectMgr::GetConditions(uint32 conditionId, std::vector<PlayerCondition const*>& out) const
 {
@@ -2563,143 +1566,6 @@ void ObjectMgr::DeleteCorpseCellData(uint32 mapid, uint32 cellid, uint32 player_
     cell_guids.corpses.erase(player_guid);
 }
 
-/**
- * @brief Loads quest relation mappings for a specific actor and role.
- *
- * @param map The destination relations map.
- * @param actor The quest actor type.
- * @param role The quest relation role.
- */
-void ObjectMgr::LoadQuestRelationsHelper(QuestRelationsMap& map, QuestActor actor, QuestRole role)
-{
-    map.clear();                                            // need for reload case
-
-    uint32 count = 0;
-
-    QueryResult* result = WorldDatabase.PQuery("SELECT `entry`, `quest` FROM `quest_relations` WHERE `actor` = %d AND `role` = %d", actor, role);
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString();
-        sLog.outErrorDb(">> Loaded 0 quest relations. DB table `quest_relations` is empty.");
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 id    = fields[0].GetUInt32();
-        uint32 quest = fields[1].GetUInt32();
-
-        if (mQuestTemplates.find(quest) == mQuestTemplates.end())
-        {
-            sLog.outErrorDb("Table `quest_relations`: Quest %u listed for entry %u does not exist.", quest, id);
-            continue;
-        }
-
-        map.insert(QuestRelationsMap::value_type(id, quest));
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString();
-    sLog.outString(">> Loaded %u %s quest %s from `quest_relations`", count, (actor == 1) ? "gameobject" : "creature", (role == 1) ? "takers" : "givers");
-}
-
-/**
- * @brief Loads quest-giver relations for gameobjects.
- */
-void ObjectMgr::LoadGameobjectQuestRelations()
-{
-    LoadQuestRelationsHelper(m_GOQuestRelations, QA_GAMEOBJECT, QR_START);
-
-    for (QuestRelationsMap::iterator itr = m_GOQuestRelations.begin(); itr != m_GOQuestRelations.end(); ++itr)
-    {
-        GameObjectInfo const* goInfo = GetGameObjectInfo(itr->first);
-        if (!goInfo)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data for nonexistent gameobject entry (%u) and existing quest %u", itr->first, itr->second);
-        }
-        else if (goInfo->type != GAMEOBJECT_TYPE_QUESTGIVER)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data gameobject entry (%u) for quest %u, but GO is not GAMEOBJECT_TYPE_QUESTGIVER", itr->first, itr->second);
-        }
-    }
-}
-
-/**
- * @brief Loads quest-completion relations for gameobjects.
- */
-void ObjectMgr::LoadGameobjectInvolvedRelations()
-{
-    LoadQuestRelationsHelper(m_GOQuestInvolvedRelations, QA_GAMEOBJECT, QR_END);
-
-    for (QuestRelationsMap::iterator itr = m_GOQuestInvolvedRelations.begin(); itr != m_GOQuestInvolvedRelations.end(); ++itr)
-    {
-        GameObjectInfo const* goInfo = GetGameObjectInfo(itr->first);
-        if (!goInfo)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data for nonexistent gameobject entry (%u) and existing quest %u", itr->first, itr->second);
-        }
-        else if (goInfo->type != GAMEOBJECT_TYPE_QUESTGIVER)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data gameobject entry (%u) for quest %u, but GO is not GAMEOBJECT_TYPE_QUESTGIVER", itr->first, itr->second);
-        }
-    }
-}
-
-/**
- * @brief Loads quest-giver relations for creatures.
- */
-void ObjectMgr::LoadCreatureQuestRelations()
-{
-    LoadQuestRelationsHelper(m_CreatureQuestRelations, QA_CREATURE, QR_START);
-
-    for (QuestRelationsMap::iterator itr = m_CreatureQuestRelations.begin(); itr != m_CreatureQuestRelations.end(); ++itr)
-    {
-        CreatureInfo const* cInfo = GetCreatureTemplate(itr->first);
-        if (!cInfo)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data for nonexistent creature entry (%u) and existing quest %u", itr->first, itr->second);
-        }
-        else if (!(cInfo->NpcFlags & UNIT_NPC_FLAG_QUESTGIVER))
-        {
-            sLog.outErrorDb("Table `quest_relations` has creature entry (%u) for quest %u, but npcflag does not include UNIT_NPC_FLAG_QUESTGIVER", itr->first, itr->second);
-        }
-    }
-}
-
-/**
- * @brief Loads quest-completion relations for creatures.
- */
-void ObjectMgr::LoadCreatureInvolvedRelations()
-{
-    LoadQuestRelationsHelper(m_CreatureQuestInvolvedRelations, QA_CREATURE, QR_END);
-
-    for (QuestRelationsMap::iterator itr = m_CreatureQuestInvolvedRelations.begin(); itr != m_CreatureQuestInvolvedRelations.end(); ++itr)
-    {
-        CreatureInfo const* cInfo = GetCreatureTemplate(itr->first);
-        if (!cInfo)
-        {
-            sLog.outErrorDb("Table `quest_relations` have data for nonexistent creature entry (%u) and existing quest %u", itr->first, itr->second);
-        }
-        else if (!(cInfo->NpcFlags & UNIT_NPC_FLAG_QUESTGIVER))
-        {
-            sLog.outErrorDb("Table `quest_relations` has creature entry (%u) for quest %u, but npcflag does not include UNIT_NPC_FLAG_QUESTGIVER", itr->first, itr->second);
-        }
-    }
-}
 
 /**
  * @brief Gets the internal locale index for a locale constant.
@@ -2762,91 +1628,6 @@ int ObjectMgr::GetOrNewIndexForLocale(LocaleConstant loc)
     return m_LocalForIndex.size() - 1;
 }
 
-/**
- * @brief Builds the set of gameobject entries that must activate for quests.
- */
-void ObjectMgr::LoadGameObjectForQuests()
-{
-    mGameObjectForQuestSet.clear();                         // need for reload case
-
-    if (!sGOStorage.GetMaxEntry())
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded 0 GameObjects for quests");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(sGOStorage.GetRecordCount());
-    uint32 count = 0;
-
-    // collect GO entries for GO that must activated
-    for (SQLStorageBase::SQLSIterator<GameObjectInfo> itr = sGOStorage.getDataBegin<GameObjectInfo>(); itr < sGOStorage.getDataEnd<GameObjectInfo>(); ++itr)
-    {
-        bar.step();
-        switch (itr->type)
-        {
-            case GAMEOBJECT_TYPE_QUESTGIVER:
-            {
-                if (m_GOQuestRelations.find(itr->id) != m_GOQuestRelations.end() ||
-                    m_GOQuestInvolvedRelations.find(itr->id) != m_GOQuestInvolvedRelations.end())
-                {
-                    mGameObjectForQuestSet.insert(itr->id);
-                    ++count;
-                }
-
-                break;
-            }
-            case GAMEOBJECT_TYPE_CHEST:
-            {
-                // scan GO chest with loot including quest items
-                uint32 loot_id = itr->GetLootId();
-
-                // always activate to quest, GO may not have loot, OR find if GO has loot for quest.
-                if (itr->chest.questId || LootTemplates_Gameobject.HaveQuestLootFor(loot_id) ||
-                    LootTemplates_Gameobject.HaveConditionalLootFor(loot_id))
-                {
-                    mGameObjectForQuestSet.insert(itr->id);
-                    ++count;
-                }
-                break;
-            }
-            case GAMEOBJECT_TYPE_GENERIC:
-            {
-                if (itr->_generic.questID)                  // quest related objects, has visual effects
-                {
-                    mGameObjectForQuestSet.insert(itr->id);
-                    ++count;
-                }
-                break;
-            }
-            case GAMEOBJECT_TYPE_SPELL_FOCUS:
-            {
-                if (itr->spellFocus.questID)                // quest related objects, has visual effect
-                {
-                    mGameObjectForQuestSet.insert(itr->id);
-                    ++count;
-                }
-                break;
-            }
-            case GAMEOBJECT_TYPE_GOOBER:
-            {
-                if (itr->goober.questId)                    // quests objects
-                {
-                    mGameObjectForQuestSet.insert(itr->id);
-                    ++count;
-                }
-                break;
-            }
-            default:
-                break;
-        }
-    }
-
-    sLog.outString(">> Loaded %u GameObjects for quests", count);
-    sLog.outString();
-}
 
 /**
  * @brief Logs a formatted script text lookup error for a string entry.
@@ -3093,52 +1874,6 @@ const char* ObjectMgr::GetMangosString(int32 entry, int locale_idx) const
     return "<error>";
 }
 
-/**
- * @brief Loads base fishing skill requirements for areas.
- */
-void ObjectMgr::LoadFishingBaseSkillLevel()
-{
-    mFishingBaseForArea.clear();                            // for reload case
-
-    uint32 count = 0;
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`,`skill` FROM `skill_fishing_base_level`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outErrorDb(">> Loaded `skill_fishing_base_level`, table is empty!");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-
-        Field* fields = result->Fetch();
-        uint32 entry  = fields[0].GetUInt32();
-        int32 skill   = fields[1].GetInt32();
-
-        AreaTableEntry const* fArea = GetAreaEntryByAreaID(entry);
-        if (!fArea)
-        {
-            sLog.outErrorDb("AreaId %u defined in `skill_fishing_base_level` does not exist", entry);
-            continue;
-        }
-
-        mFishingBaseForArea[entry] = skill;
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u areas for fishing base skill level", count);
-    sLog.outString();
-}
 
 // Check if a player meets condition conditionId
 bool ObjectMgr::IsPlayerMeetToCondition(uint16 conditionId, Player const* pPlayer, Map const* map, WorldObject const* source, ConditionSource conditionSourceType) const
@@ -4829,33 +3564,6 @@ bool LoadMangosStrings(DatabaseType& db, char const* table, int32 start_value, i
     return sObjectMgr.LoadMangosStrings(db, table, start_value, end_value, extra_content);
 }
 
-/**
- * @brief Loads creature template spell assignments and validates their spell ids.
- */
-void ObjectMgr::LoadCreatureTemplateSpells()
-{
-    sCreatureTemplateSpellsStorage.Load();
-
-    for (SQLStorageBase::SQLSIterator<CreatureTemplateSpells> itr = sCreatureTemplateSpellsStorage.getDataBegin<CreatureTemplateSpells>(); itr < sCreatureTemplateSpellsStorage.getDataEnd<CreatureTemplateSpells>(); ++itr)
-    {
-        if (!sCreatureStorage.LookupEntry<CreatureInfo>(itr->entry))
-        {
-            sLog.outErrorDb("LoadCreatureTemplateSpells: Spells found for creature entry %u, but creature does not exist, skipping", itr->entry);
-            sCreatureTemplateSpellsStorage.EraseEntry(itr->entry);
-        }
-        for (uint8 i = 0; i < CREATURE_MAX_SPELLS; ++i)
-        {
-            if (itr->spells[i] && !sSpellStore.LookupEntry(itr->spells[i]))
-            {
-                sLog.outErrorDb("LoadCreatureTemplateSpells: Spells found for creature entry %u, assigned spell %u does not exist, set to 0", itr->entry, itr->spells[i]);
-                const_cast<CreatureTemplateSpells*>(*itr)->spells[i] = 0;
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u creature_template_spells definitions", sCreatureTemplateSpellsStorage.GetRecordCount());
-    sLog.outString();
-}
 
 /**
  * @brief Retrieves a creature template from the global creature store.
